requirements.txt
langchain
langchain-community
neo4j
langchain-experimental
youtube-transcript-api
streamlit
python-dotenv
langchain-google-genai
json-repair
langchain_groq